feat: modernise frontend
