+++ client/.env.example
@@ -1,4 +1,3 @@
-
 # WEBSERVER DETAILS
 REACT_APP_API_URL = YOUR_BACKEND_API_URL_STRING
 REACT_APP_TEST_API_URL = http://YOUR_BACKEND_TEST_API_URL_STRING:1120
