+++ client/src/pages/components/navbar.js
@@ -1,8 +1,6 @@
-// Modules
 import { Link } from "react-router-dom";
 
-export default function Navbar(props) {
-
+export default function Navbar(_props) {
   return (
     <div className="flex flex-row items-center justify-between w-screen py-3 px-5 md:px-20 bg-gray-300">
       <div className="flex flex-row items-center justify-center">
@@ -14,5 +12,4 @@ export default function Navbar(props) {
       </div>
     </div>
   )
-
-}
+};
