+++ client/src/styles/custom.css
@@ -1,4 +1,3 @@
-
 .font-inter {
   font-family: "Inter", sans-serif;
 }
